refactor(governance): one place reads the constraint codes back
The identity links and the key-to-bill mapping both hold their rules as
database constraints, so both have to recognise a raw SQLSTATE through
whichever shape Prisma wrapped it in. Moves the reader out of the match
engine's error file and adds the check-violation case beside it, rather
than copying the message-parsing subtlety into a second place where the
two could drift.

No behaviour change: the single caller now imports it from the shared
module.
