+++ platform/app/ee/governance/services/identityMatch.errors.ts
@@ -2,10 +2,9 @@
 
 /**
  * What the match engine refuses to do, said in words the reviewer can act on
- * (ADR-128 §12) — plus the domain's shared reading of Postgres refusing a
- * duplicate, which the discovery repository leans on too.
+ * (ADR-128 §12).
  *
- * The error classes are races rather than mistakes: a review queue is read by
+ * Both of these are races rather than mistakes: a review queue is read by
  * people, and between reading it and clicking, the world moves. The whole point
  * of naming them is that the loser of such a race should be told what happened,
  * not handed a trace id.
@@ -84,49 +83,4 @@ export class IdentityErasedError extends HandledError {
     );
     this.name = "IdentityErasedError";
   }
-}
-
-/** Postgres' unique-violation code. */
-const UNIQUE_VIOLATION = "23505";
-
-/**
- * Prisma's own code for the same refusal: the client wraps a unique violation
- * as a `PrismaClientKnownRequestError` with `code: "P2002"` and buries the
- * SQLSTATE underneath. Which unique rule tripped is the caller's knowledge —
- * the one-open-link index for `IdentityMatch` writes, the (organization,
- * provider, actor) key for discovered-person writes.
- */
-const PRISMA_UNIQUE_VIOLATION = "P2002";
-
-/**
- * Whether a thrown value is Postgres refusing a duplicate under any unique
- * rule.
- *
- * Reads the code off whichever shape carried it: Prisma's `P2002` wrapper, a
- * raw driver error's SQLSTATE in `code` or `meta.code`, or — for the wrapper
- * shapes whose SQLSTATE lives only in the message — a word-bounded match on
- * the message. Never a `String(err).includes` over the whole message: that
- * would also match an error whose *payload* happened to contain the digits.
- */
-export function isUniqueViolation(error: unknown): boolean {
-  if (typeof error !== "object" || error === null) return false;
-  const code = (error as { code?: unknown }).code;
-  if (code === UNIQUE_VIOLATION || code === PRISMA_UNIQUE_VIOLATION)
-    return true;
-  const meta = (error as { meta?: { code?: unknown } }).meta;
-  if (meta?.code === UNIQUE_VIOLATION) return true;
-  const message = (error as { message?: unknown }).message;
-  return (
-    typeof message === "string" &&
-    new RegExp(`\\b${UNIQUE_VIOLATION}\\b`).test(message)
-  );
-}
-
-/**
- * Whether a thrown value is the one-open-link index refusing a second open
- * link. The same refusal shape as any unique violation; the name records
- * which rule the caller believes it tripped (`IdentityMatch` has no other).
- */
-export function isOpenLinkViolation(error: unknown): boolean {
-  return isUniqueViolation(error);
 }
